Sorted extensions and views by files and folders
