+++ Habit Helper/Controllers/CongratulationScreen/CongratulationScreen.swift
@@ -1,10 +1,3 @@
-//
-//  CongratulationScreen.swift
-//  Habit Helper
-//
-//  Created by Анастасия Гаранович on 21.10.2021.
-//
-
 import UIKit
 import iOSTools
