+++ Habit Helper/Extensions/UIButton.swift
@@ -1,0 +1,8 @@
+import UIKit
+
+extension UIButton {
+    func setButtonBorderColor() {
+        layer.borderWidth = 2.5
+        layer.borderColor = UIColor(named: "border")?.cgColor
+    }
+}
