+++ Habit Helper/Extensions/UICollectionViewCell.swift
@@ -1,0 +1,8 @@
+import UIKit
+
+extension UICollectionViewCell {
+    func setCellBorderColor(color: String) {
+        layer.borderColor = UIColor(named: color)?.cgColor
+        layer.borderWidth = 2.5
+    }
+}
