--- Habit Helper/Storyboards/CongratulationScreen/CongratulationScreen.storyboard
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="19162" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="19144"/>
-        <capability name="Named colors" minToolsVersion="9.0"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="System colors in document resources" minToolsVersion="11.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <customFonts key="customFonts">
-        <array key="Nunito-Regular.ttf">
-            <string>Nunito-Regular</string>
-        </array>
-    </customFonts>
-    <scenes>
-        <!--View Controller-->
-        <scene sceneID="s0d-6b-0kx">
-            <objects>
-                <viewController id="Y6W-OH-hqX" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="5EZ-qb-Rvc">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="logo" translatesAutoresizingMaskIntoConstraints="NO" id="DJG-aa-iYd">
-                                <rect key="frame" x="82" y="244" width="250" height="250"/>
-                                <constraints>
-                                    <constraint firstAttribute="height" constant="250" id="LNk-aW-chB"/>
-                                    <constraint firstAttribute="width" constant="250" id="Vkz-M7-hnj"/>
-                                </constraints>
-                            </imageView>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Congratulation with new life" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="S4k-aJ-aCq">
-                                <rect key="frame" x="92" y="598" width="230" height="25"/>
-                                <fontDescription key="fontDescription" name="Nunito-Regular" family="Nunito" pointSize="18"/>
-                                <color key="textColor" systemColor="systemGray3Color"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Well done!" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="gmj-w0-AKi">
-                                <rect key="frame" x="128.5" y="539" width="157" height="44"/>
-                                <fontDescription key="fontDescription" name="Nunito-Regular" family="Nunito" pointSize="32"/>
-                                <color key="textColor" name="buttons"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                        </subviews>
-                        <viewLayoutGuide key="safeArea" id="vDu-zF-Fre"/>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                        <constraints>
-                            <constraint firstItem="S4k-aJ-aCq" firstAttribute="top" secondItem="gmj-w0-AKi" secondAttribute="bottom" constant="15" id="Chs-2Z-ccX"/>
-                            <constraint firstItem="DJG-aa-iYd" firstAttribute="centerX" secondItem="5EZ-qb-Rvc" secondAttribute="centerX" id="EvH-mS-867"/>
-                            <constraint firstItem="gmj-w0-AKi" firstAttribute="top" secondItem="DJG-aa-iYd" secondAttribute="bottom" constant="45" id="ekD-Ba-QzK"/>
-                            <constraint firstItem="S4k-aJ-aCq" firstAttribute="centerX" secondItem="5EZ-qb-Rvc" secondAttribute="centerX" id="lNW-Pi-1rB"/>
-                            <constraint firstItem="gmj-w0-AKi" firstAttribute="centerX" secondItem="5EZ-qb-Rvc" secondAttribute="centerX" id="rHO-Zq-UnD"/>
-                            <constraint firstItem="DJG-aa-iYd" firstAttribute="top" secondItem="vDu-zF-Fre" secondAttribute="top" constant="200" id="wfq-FN-RuY"/>
-                        </constraints>
-                    </view>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="Ief-a0-LHa" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="116" y="107"/>
-        </scene>
-    </scenes>
-    <resources>
-        <image name="logo" width="1024" height="1024"/>
-        <namedColor name="buttons">
-            <color red="0.50199997425079346" green="0.84700000286102295" blue="0.60799998044967651" alpha="1" colorSpace="custom" customColorSpace="displayP3"/>
-        </namedColor>
-        <systemColor name="systemBackgroundColor">
-            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-        </systemColor>
-        <systemColor name="systemGray3Color">
-            <color red="0.7803921568627451" green="0.7803921568627451" blue="0.80000000000000004" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-        </systemColor>
-    </resources>
-</document>
